fix(ui): dispatch uGUI button clicks once
