+++ unity-package/Editor/Util/UI.UGUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e79bb19f358e4d898b3dc8dbe5f2d810
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
