+++ unity-package/Tests/Editor/UIInteractionTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5b70c20a1cf04dde99e9d768cab8df37
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
